requirements.txt
python-telegram-bot==21.6
httpx>=0.27
python-dotenv